Always return the output from ref-update hook
Change I5905c44c7 moved the invocation of the ref-update hook out of
the ReceiveCommits implementation and into CommitValidators. Since
then, the output of the ref-update hook is not sent back to the git
client except in the case of an error or rejection.

Update the handling of the ref-update hook in the plugin so that its
output is returned back to the caller as a list of validation messages.

Each line of the hook's output is sent as a separate validation message
instance. This is to make the output properly appear as multiple lines
on the client, for example:

    remote: Processing changes: new: 1, refs: 1, done
    remote: message from ref-update hook line 1
    remote: message from ref-update hook line 2
    remote: message from ref-update hook line 3

Change-Id: I88af8a3c690cc226214208c3c93f552f054ccae0
Reported-By: Robert Niemi <robert.den.klurige@googlemail.com>