ref-update: Let hooks distinguish magic ref pushes
Prior to this changeset, the logic about what happens with
ref-update when someone pushes to refs/for/<branch> was broken.
Further, the only way to tell if the push was a review to
refs/changes/<branch> was indistinguishable from there being
no parent.

Instead, let's simply add a new argument that allows the hook
to make the appropriate decision itself and not overload
semantics with the previous oldrev.

Change-Id: I24184d47234123450a60946af159e7c9f8375057