Rework ref-update hook to use RefOperationValidationListener
Instead of being invoked on every commit received, ref-update is now
invoked before every ref update operation is finalized.

The previous behaviour of the ref-update hook is moved to a new hook
named commit-received. A new parameter '--cmdref' is added, and the
special handling of 'refs/for' and 'refs/changes' is removed.

Change-Id: I24d8580aef463f8bfb078d90a4e9e2ce6e8b9b93