Update patch set 2
Patch Set 2:

> only reason i did ref-operate in the previous change was for
 > backwards compatibility concerns. sounds like we don't care too
 > much.
 > 

If we were doing this on the stable branch it would be a concern.  Not so much on master, as long as we document it clearly in the release notes.

 > also not sure why i lost authorship for renaming, but obviously i'm
 > +1 on this otherwise.

Corrected the author name on the latest PS.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>