Fix hook configuration names to match documentation
Most hooks in the documentation are named with "Hook" suffix, but
in the code the "Hook" suffix was omitted. This prevented the hook
from being found when the administrator configured it with the
documented name.

Fix the values used, to match the documented values.

Also update the documentation to add the "Hook" suffix where it was
missing.

Bug: Issue 6508
Change-Id: Icdfdaaa2a4b89ed636e9165fc073234fb79abd2b