RFC: Add username arguments to hooks that take AccountInfo
When an argument is added from an AccountInfo object, it gets
formatted as:

 Name (Email)

or if the account does not have an email address:

 Name

where "Name" can be "Anonymous Coward" if the name was not set.

From this information hooks cannot easily get the account username.

For the arguments that take an AccountInfo, also add an argument for
the username. So for example if the hook takes the argument:

  --change-owner Name (Email)

it will now also have:

  --change-owner-username username

Change-Id: Ic38358464122c11e7e8f7616fbafa14bf71aef99
TODO: add documentation of the new ...-username arguments