RefUpdate: Protect against null ObjectId in ReceiveCommand
RefUpdate's onRefOperation method is invoked by RefOperationValidators
for every registered RefOperationValidationListener.

If one of those listeners passes a RefReceivedEvent that contains an
instance of ReceiveCommand that was initialized with null old or
new ObjectId, this results in a NullPointerException when the ObjectId
is dereferenced to get the name.

Protect against such misbehaving validators by translating any null
instance to ObjectId.zeroId().

Change-Id: Ib9399c2351af20b23f8c4481c7b067f3a4dcb688