Add documentation to help with debugging hook problems
A common problem for users of the plugin is that they are not sure
if their hooks are actually being executed, and what their output
is.

Such information is logged by the plugin, but only at debug level.

Add a section in the documentation showing how to enable the logging.

Change-Id: I6e040f958ce4503f5cc0468ebe09466914d4a7eb