Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I wonder should a 2.15 branch be created now for gerrit 2.15?

No. We only need to create a stable-2.15 branch on the plugin if there are fixes that are needed.

The changes here are already included in the master branch revision that is used in core gerrit stable-2.15.

Patch-set: 1
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>