Add README
Add a README.md file in the repository root with the same content as
the "About" page.

Gitiles will render the README by default when the project is viewed
on the googlesource site [1] and this will make the configuration
documentation and list of supported hooks more easily accessible for
users who have not installed the plugin yet in their Gerrit site.

[1] https://gerrit.googlesource.com/plugins/hooks

Change-Id: Ib2ecf818cbabaca8bfddf58e12cce99842a4f038