More logging improvements to help tracking down hook problems
Add more informational logs:

- The configured hooks path.
- The resolved path for each hook.

Each of these is only logged once (the latter once per hook) so logging
at info level will not be obtrusive.

Log the path of the hook file if it does not exist. This is logged at
debug level because it will be emitted on every attempt to execute the
hook.

Change-Id: I1fe6f860346d73648283178b58982180d2c35ce9