Add submit hook
The submit hook is invoked synchronously when a change is submitted. If
it returns a non-zero exit status, a MergeValidationException is thrown
and the submit is prevented.

This adds back the ability to block submit by a hook. Previously this
was possible with the ref-update hook until its purpose was changed to
only block ref updates such as branch creation, deletion, or fast-forward
by direct push.

TODO: Documentation

Change-Id: Ie4efb90df645ecac01638b23305dc2ffb547192e