HookQueue: Lazily instantiate the queue
HookQueue is a LifecycleListener, and creates the queue in its onStart
method. This means that the queue is created when the plugin is loaded.

In the case where the plugin is reloaded, this occurs before the old
version of the plugin has stopped, and the old plugin's queue (and
metrics) still exist. Then, when the new plugin's queue is created,
creating its metrics fails because they already exist.

Change it to lazily create the queue the first time a hook is executed.

Change-Id: Ife59e26cba4f0a6d7338a0850a64cccb8ba31252