Resolve relative hooks.path against $site_path, not cwd of the process
When hooks.path was given as relative path it got resolved against the
current working directory of the Gerrit process. While it may often be
the case that the current working directory of Gerrit process is the
$site_path, it is not guaranteed and, depending how Gerrit is started,
it may be some other directory. Most, if not all, of the relative paths
in gerrit.config are resolved against the $site_path and we should do
the same for the hooks.path.

Release-Notes: Resolve relative hooks.path against $site_path
Change-Id: Ie72c0657393e1dfe078155e27edbd4c0412892e6